requirements.txt
cbpro==1.1.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
configparser==5.0.1
Flask==1.1.2
idna==2.10
influxdb==5.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
lbcapi3==1.0.0
MarkupSafe==1.1.1
msgpack==1.0.2
oandapyV20==0.6.3
paxful-api==1.0.1
pymongo==3.5.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.0
simplejson==3.17.2
six==1.10.0
sortedcontainers==2.3.0
tabulate==0.8.7
termcolor==1.1.0
urllib3==1.26.2
websocket-client==0.40.0
Werkzeug==1.0.1
